allow opt-out for the assignee of an issue
if configured and set to the respective value, the initial merge request
will simply be ignored.

NOTE: this check is only performed during the actual webhook trigger from
Jira; it is not checked again after the configured MergeDelay.
Setting the custom field after the issue is closed has no effect and will
most likely be too late anyways.