handle local file paths aswell
for windows, it might look like a scheme...except that the "scheme" is a
drive letter.
for linux and relative paths, we might not have a colon at all.

just assume those two cases to be paths and only take the last part as
relative path name when referring to the repo in any place.