allow certain branches to be ignored for automerging
branches might be pushed to a namespace such as "private/" or "testing/"
for temporary use, but stick around until the point where the merge would
take place. to avoid those breaking an automatic merge, lets allow them to
be excluded in first place.

this only filters them during a non-exact match based on the issue key;
merging an exactly specified branch should always be possible (regardless
of whether it is considered an ignored branch or not)

Fixes #7